MGMT-18103: Modify existing go code to support api and worker servers
Note: this PR doesn't contain any changes in Makefile, Dockerfile, and
deployment scripts.

The main idea of this PR is to separate the existing server to api,
worker, and periodic.  It is based on the flow in TaskManager.  All
logic behind the channel listeneres is now handled by the worker server.
The api server is the one listening to web interface and executes the
API commands.
The periodic server periodically wakes up and executes periodic tasks.

Currently some of the code is duplicated.  Eventually the duplicated old code will
be removed.