EDM-1445: Private GIT repo doesn’t allow authentication with a token
This is a regression since the service started using API entities
instead of store/model entities.  The reason for that is that the
function "ToApiResource" of model.Respository hides sensitive data.
Therefore, this data cannot used internally by the service.
The solution was to move the hiding logic to "transport", before the
data is sent to the client.